refactor: ship the lucide resolver as its own module
An external build that imports every icon explicitly — a desk island,
say — needs the `~icons/lucide/<name>` resolver and nothing else, but the
only entry available drags in unplugin-auto-import and
unplugin-vue-components for an auto-import DX it will never use.

The resolver and the icon map move to `vite/lucideIconsPlugin.js`, exposed
as `frappe-ui/vite/lucideIconsPlugin`. `vite/lucideIcons.js` keeps
composing it with the unplugins, so the existing entry is unchanged.
